==> 03_Baseline_Model.ipynb <==
import pandas as pd
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import StandardScaler
from sklearn.linear_model import LinearRegression
from sklearn.metrics import mean_squared_error
from sklearn.metrics import r2_score
from sklearn.preprocessing import LabelEncoder

import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

df = pd.read_csv('data/kickstarter_projects.csv')
df.head()

# convert columns to lowercase
df.columns = df.columns.str.lower()

# drop 'ID' and 'Name' columns
df = df.drop(columns=['id', 'name'])
df.columns

# print correlation of variables
#df[['category', 'subcategory', 'country', 'goal']].corr()

df.describe()

# define feature and target variable
X = df[['goal']]
y = df['pledged']

X.shape

# fit linear regression model
lin_reg = LinearRegression()
lin_reg.fit(X, y)

# calculate r-squared 
y_hat = lin_reg.predict(X)
print("R-squared:", r2_score(y, y_hat))

# plot data with regression line
plt.scatter(X, y)
plt.plot(X, y_hat, '-', color='orange', linewidth=2)
plt.ylabel('pledged')
plt.xlabel('goal')

# define features and target variable
X = df[['goal']]
y = df['pledged']

# apply logarithmic transformation to 'goal' and 'pledged' columns
X['goal'] = np.log1p(X['goal'])  # log1p is used to handle zero values
y = np.log1p(y)

# split data into training and testing sets
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)

# scale features
scaler = StandardScaler()
X_train_scaled = scaler.fit_transform(X_train)
X_test_scaled = scaler.transform(X_test)

# train linear regression model
model = LinearRegression()
model.fit(X_train_scaled, y_train)

# examine distributions after log transformation
for col in ['goal']:
    sns.histplot(X[col], kde=True)
    plt.title(f'Log Transformed Distribution of {col}')
    plt.show()
    
sns.histplot(y, kde=True)
plt.title('Log Transformed Distribution of Pledged')
plt.show()

# make predictions
y_pred = model.predict(X_test_scaled)

# evaluate model
mse = mean_squared_error(y_test, y_pred)
print(f'Mean Squared Error: {mse}')

# inverse transform predictions to original scale
y_pred_original_scale = np.expm1(y_pred)  # inverse of log1p is expm1
y_test_original_scale = np.expm1(y_test)

# evaluating the model with original target values
mse_original_scale = mean_squared_error(y_test_original_scale, y_pred_original_scale)
print(f'Mean Squared Error (Original Scale): {mse_original_scale}')
