==> 05_Model_KNN_State_Exploration.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns

from sklearn.metrics import confusion_matrix, accuracy_score, recall_score
from sklearn.neighbors import KNeighborsClassifier
from sklearn.model_selection import train_test_split

df = pd.read_csv("data/binary_state_kickstarter_projects.csv")
df.head()

# correlation heatmap 
correlations = df.corr(numeric_only=True)
mask = np.triu(correlations)
sns.heatmap(correlations , vmax=1, vmin=-1, annot=True, mask=mask, cmap="YlGnBu");

# defining target and other variables
X = df[['duration']]
y = df['state']

# split data into train and test set
X_train, X_test, y_train, y_test = train_test_split(X, y, random_state=1)

# train model
knn = KNeighborsClassifier()
knn.fit(X_train, y_train)

# predict on test set
y_pred = knn.predict(X_test)

# print accuracy score 
print("Accuracy:", accuracy_score(y_test, y_pred).round(2))
print("Recall:", recall_score(y_test, y_pred, average='weighted').round(2))
print("-----"*10)

# print confusion matrix
sns.heatmap(confusion_matrix(y_test, y_pred), annot=True, cmap='YlGn');

## Using Categorical Variables Country and Category

# defining target and other variables
X = df[['country_encoded','category_encoded']]
y = df['state']

# split data into train and test set
X_train, X_test, y_train, y_test = train_test_split(X, y, random_state=1)

# train model
knn = KNeighborsClassifier()
knn.fit(X_train, y_train)

# predict on test set
y_pred = knn.predict(X_test)

# print accuracy score 
print("Accuracy:", accuracy_score(y_test, y_pred).round(2))
print("Recall:", recall_score(y_test, y_pred, average='weighted').round(2))
print("-----"*10)

# print confusion matrix
sns.heatmap(confusion_matrix(y_test, y_pred), annot=True, cmap='YlGn');

## Using Continuous Variable Goal

# defining target and other variables
X = df[['goal']]
y = df['state']

# split data into train and test set
X_train, X_test, y_train, y_test = train_test_split(X, y, random_state=1)

# train model
knn = KNeighborsClassifier()
knn.fit(X_train, y_train)

# predict on test set
y_pred = knn.predict(X_test)

# print accuracy score 
print("Accuracy:", accuracy_score(y_test, y_pred).round(2))
print("Recall:", recall_score(y_test, y_pred, average='weighted').round(2))
print("-----"*10)

# print confusion matrix
sns.heatmap(confusion_matrix(y_test, y_pred), annot=True, cmap='YlGn');

## Mixing it together

# defining target and other variables
X = df[['goal','country_encoded','duration','category_encoded']]
y = df['state']

# split data into train and test set
X_train, X_test, y_train, y_test = train_test_split(X, y, random_state=1)

# train model
knn = KNeighborsClassifier()
knn.fit(X_train, y_train)

# predict on test set
y_pred = knn.predict(X_test)
s
# print accuracy score 
print("Accuracy:", accuracy_score(y_test, y_pred).round(2))
print("Recall:", recall_score(y_test, y_pred, average='weighted').round(2))
print("-----"*10)

# print confusion matrix
sns.heatmap(confusion_matrix(y_test, y_pred), annot=True, cmap='YlGn');
