==> 06_Model_Binary_Logistic.ipynb <==
import numpy as np
import pandas as pd
pd.set_option('display.float_format', lambda x: '%.3f' % x) #Set the pandas display format to suppress scientific notation
import matplotlib.pyplot as plt
import seaborn as sns

from sklearn.metrics import confusion_matrix, accuracy_score, recall_score
from sklearn.linear_model import LogisticRegression
from sklearn.model_selection import train_test_split

df = pd.read_csv("data/binary_state_kickstarter_projects.csv")
df.head()

# scaling nummerical data

from sklearn.preprocessing import StandardScaler

numerical = df[['goal', 'duration', 'category_encoded', 'country_encoded']]
numerical2 = df[['backers', 'pledged', 'goal', 'duration']]
numerical3 = df[['goal', 'duration']]

scaler = StandardScaler()

# import encoder
from sklearn.preprocessing import LabelEncoder

# defining X and y
X = scaler.fit_transform(numerical3)
y = df['state']

pd.DataFrame(X).describe()

from sklearn.model_selection import train_test_split
from sklearn.linear_model import LogisticRegression

# splitting  dataset into train set and test set
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size = 0.2, random_state = 21)

# train a Logistic Regression
algorithm = LogisticRegression(random_state = 22)
algorithm.fit(X_train, y_train)

from sklearn.model_selection import cross_val_score

# cross validation
accuracies = cross_val_score(estimator = algorithm, X = X_train, y = y_train, cv = 5)

print(accuracies.mean())
print(accuracies.std())

print(algorithm.score(X_test, y_test))

# make prediction
y_pred = algorithm.predict(X_test)

from sklearn.metrics import classification_report

# classification report
cm = classification_report(y_test, y_pred)
print(cm)

# transform nummerical labels to original state types
true_projects = y_test
predicted_projects = y_pred

# create df and cross-tabulation
df = pd.DataFrame({'true_projects': true_projects, 'predicted_projects': predicted_projects}) 
pd.crosstab(df.true_projects, df.predicted_projects)
