==> 07_Model_KMeans.ipynb <==
import numpy as np
import pandas as pd

import matplotlib.pyplot as plt
import seaborn as sns

# set display format to suppress scientific notation
pd.set_option('display.float_format', lambda x: '%.3f' % x)

df = pd.read_csv("data/binary_state_kickstarter_projects.csv")
df.head()

columns_to_drop = ['category','subcategory','country','launched', 'deadline']
df.drop(columns=columns_to_drop, inplace=True, axis=1)
df.head()

x = df['state'].value_counts()
x_male_percentage = int(x[1]/len(df) * 100)
x_female_percentage = int(x[0]/len(df) * 100)
print(f'Successful: {x_male_percentage}% \nFailed: {x_female_percentage}%')

# plot goal and pledged
ax = sns.scatterplot(data=df, x='duration',y='goal', hue='state')
#plt.xscale('log')
plt.yscale('log')

# plot goal and pledged
ax = sns.scatterplot(data=df, x='launch_month',y='goal', hue='state')
#plt.xscale('log')
plt.yscale('log')

# plot goal and pledged
ax = sns.scatterplot(data=df, x='category_encoded',y='goal', hue='state')
#plt.xscale('log')
plt.yscale('log')

from sklearn.cluster import KMeans
 
X = df[['duration']]
y = ['state']

algorithm = KMeans(n_clusters=2)
y_means = algorithm.fit_predict(X)

df_cluster = pd.Series(y_means, name='Cluster')
df_clustered = pd.concat([df, df_cluster], axis=1)
df_clustered

from sklearn.metrics import confusion_matrix, accuracy_score, recall_score

# print accuracy score 
print("Accuracy:", accuracy_score(df_clustered['state'], df_clustered['Cluster']).round(2))

ax = sns.scatterplot(data=df_clustered, x='pledged_log',y='goal_log', hue='Cluster')
