==> 08_Visualizations.ipynb <==
import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt
import missingno as msno

df = pd.read_csv("data/binary_state_kickstarter_projects.csv") 
df.head()

# function to set the plot style
def set_plot_style():
    # set background color of plot
    plt.gca().set_facecolor('#002845')
    plt.gcf().set_facecolor('#002845')
    # set title and axis labels to white color
    plt.title('Title', color='white')
    plt.xlabel('X Label', color='white')
    plt.ylabel('Y Label', color='white')
    # set color of tick labels
    plt.xticks(color='white')
    plt.yticks(color='white')

# Category

# group by 'category' and calculate mean of 'state'
success_rate = df.groupby('category', sort=True)['state'].count().sort_values(ascending=False).reset_index()
success_rate.columns = ['category', 'success_rate']

# plot with Seaborn
plt.figure(figsize=(10, 6))
sns.barplot(y='category', x='success_rate', data=success_rate)
plt.title('Count per Category')
plt.ylabel('Count')
plt.xlabel('Category')
plt.show()

# group by 'category' and calculate mean of 'state'
success_rate = df.groupby('category')['state'].mean().sort_values(ascending=False).reset_index()
success_rate.columns = ['category', 'success_rate']

# plot with Seaborn
plt.figure(figsize=(10, 6))
sns.barplot(y='category', x='success_rate', data=success_rate)
plt.title('Avg. Success Rate per Category')
plt.ylabel('Avg. Success Rate')
plt.xlabel('Category')
plt.show()

# group by 'category' and calculate mean of 'state'
success_rate = df.groupby('category')['goal'].median().sort_values(ascending=False).reset_index()
success_rate.columns = ['category', 'success_rate']

# plot with Seaborn
plt.figure(figsize=(10, 6))
sns.barplot(y='category', x='success_rate', data=success_rate)
plt.title('Median Goal per Category')
plt.ylabel('Median Goal')
plt.xlabel('Category')
plt.show()

# group by 'category' and calculate mean of 'state'
success_rate = df.groupby('category')['pledged'].median().sort_values(ascending=False).reset_index()
success_rate.columns = ['category', 'success_rate']

# plot with Seaborn
plt.figure(figsize=(10, 6))
sns.barplot(y='category', x='success_rate', data=success_rate)
plt.title('Median Pledged per Category')
plt.ylabel('Median Pledged')
plt.xlabel('Category')
plt.show()

# Country

# group by 'country' and calculate mean of 'state'
success_rate = df.groupby('country')['state'].mean().sort_values(ascending=False).reset_index()
success_rate.columns = ['country', 'success_rate']

# plot with Seaborn
plt.figure(figsize=(10, 6))
sns.barplot(y='country', x='success_rate', data=success_rate)
plt.title('Avg. Success Rate per Country')
plt.ylabel('Avg. Success Rate')
plt.xlabel('Country')
plt.show()

# group by 'country' and calculate mean of 'state'
success_rate = df.groupby('country')['goal'].median().sort_values(ascending=False).reset_index()
success_rate.columns = ['country', 'success_rate']

# plot with Seaborn
plt.figure(figsize=(10, 6))
sns.barplot(y='country', x='success_rate', data=success_rate)
plt.title('Median Goal per Country')
plt.ylabel('Median Goal')
plt.xlabel('Country')
plt.show()

# group by 'country' and calculate mean of 'state'
success_rate = df.groupby('country')['pledged'].median().sort_values(ascending=False).reset_index()
success_rate.columns = ['country', 'success_rate']

# plot with Seaborn
plt.figure(figsize=(10, 6))
sns.barplot(y='country', x='success_rate', data=success_rate)
plt.title('Median Pledged per Country')
plt.ylabel('Median Pledged')
plt.xlabel('Country')
plt.show()
